# [FIX] mrp: optimize _compute_used_in_bom_count